test(e2e): share watch component lookup
